// packages/twenty-server/src/engine/core-modules/onboarding/enums/onboarding-status.enum.ts
export enum OnboardingStatus {
  PLAN_REQUIRED = 'PLAN_REQUIRED',
  WORKSPACE_ACTIVATION = 'WORKSPACE_ACTIVATION',
  PROFILE_CREATION = 'PROFILE_CREATION',
  CONNECT_LINKEDIN = 'CONNECT_LINKEDIN',
  INSTALL_APP = 'INSTALL_APP',
  SYNC_EMAIL = 'SYNC_EMAIL',
  INVITE_TEAM = 'INVITE_TEAM',
  COMPLETED = 'COMPLETED',
}